Verify: move implementation from server to proto
This allows verification to be checked client side, and this
functionaliy makes more sense as a method.
